// apps/monorepo-lite/apps/api/src/modules/auth/auth.service.ts
export type RegisterResult = {
  user: User
  session: Session
  organization: {
    id: string
    name: string
    slug: string
  } | null
}

export type LoginResult = {
  user: User
  session: Session
}

export type MeResult = {
  user: User
  session: Session
  organizations: {
    id: string
    name: string
    slug: string
    role: string
  }[]
}